requirements.txt
flake8 
python-chess
selenium
beautifulsoup4

